requirements.txt
# Стабильные версии для Python 3.13
aiogram==3.4.1
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-dotenv==1.0.0
asyncpg==0.28.0
python-multipart==0.0.6

# Дополнительно (если понадобится)
aiofiles==23.1.0
